make sure the seeds file runs only in test mode

+++ db/seeds.rb
@@ -7,22 +7,24 @@
 #   Mayor.create(:name => 'Daley', :city => cities.first)
 
 # run this file with 'bin/rake db:seed RAILS_ENV=test'
+if Rails.env.test?
 
-Task.delete_all
-Project.delete_all
-User.delete_all
+  Task.delete_all
+  Project.delete_all
+  User.delete_all
 
-user1 = User.create!(:password => "password", :email => "marc@email.com")
-user1.active = true
-user1.save
-user2 = User.create!(:password => "password", :email => "hacker@lulz.com")
-user2.active = true
-user2.save
+  user1 = User.create!(:password => "password", :email => "marc@email.com")
+  user1.active = true
+  user1.save
+  user2 = User.create!(:password => "password", :email => "hacker@lulz.com")
+  user2.active = true
+  user2.save
 
-project1 = Project.new(:name => "Project1")
-project1.users << user1
-project1.save()
+  project1 = Project.new(:name => "Project1")
+  project1.users << user1
+  project1.save()
 
+end
